adding support for ios app

+++ src/main/java/com/reporting/ExtentReports/ExtentTestManager.java
@@ -37,16 +37,16 @@ public class ExtentTestManager {
     static ExtentReports extent = ExtentManager.getReporter();
 
     public static synchronized ExtentTest getTest() {
-        return (ExtentTest) extentTestMap.get((int) (long) (Thread.currentThread().getId()));
+        return (ExtentTest) extentTestMap.get((int) Thread.currentThread().getId());
     }
 
     public static synchronized void endTest() {
-        extent.endTest((ExtentTest) extentTestMap.get((int) (long) (Thread.currentThread().getId())));
+        extent.endTest((ExtentTest) extentTestMap.get((int) Thread.currentThread().getId()));
     }
 
     public static synchronized ExtentTest startTest(String testName, String desc) {
         ExtentTest test = extent.startTest(testName, desc);
-        extentTestMap.put((int) (long) (Thread.currentThread().getId()), test);
+        extentTestMap.put((int) Thread.currentThread().getId(), test);
         return test;
     }
 }
